feat: write endpoint for device settings
