+++ api/tests/test_v1_endpoints.py
@@ -1,5 +1,7 @@
-import json
-from os import path
+"""
+Tests for V1 API endpoints.
+"""
+import os
 from pathlib import Path
 from unittest import mock
 
@@ -11,161 +13,8 @@ from rest_framework.test import APIClient
 from unittest_parametrize import ParametrizedTestCase, parametrize
 
 from anthias_app.models import Asset
+from api.tests.test_common import ASSET_CREATION_DATA
 from settings import settings as anthias_settings
-
-ASSET_LIST_V1_1_URL = reverse('api:asset_list_v1_1')
-ASSET_CREATION_DATA = {
-    'name': 'Anthias',
-    'uri': 'https://anthias.screenly.io',
-    'start_date': '2019-08-24T14:15:22Z',
-    'end_date': '2029-08-24T14:15:22Z',
-    'duration': 20,
-    'mimetype': 'webpage',
-    'is_enabled': 0,
-    'nocache': 0,
-    'play_order': 0,
-    'skip_asset_check': 0
-}
-ASSET_UPDATE_DATA_V1_2 = {
-    'name': 'Anthias',
-    'uri': 'https://anthias.screenly.io',
-    'start_date': '2019-08-24T14:15:22Z',
-    'end_date': '2029-08-24T14:15:22Z',
-    'duration': '15',
-    'mimetype': 'webpage',
-    'is_enabled': 1,
-    'nocache': 0,
-    'play_order': 0,
-    'skip_asset_check': 0
-}
-ASSET_UPDATE_DATA_V2 = {
-    **ASSET_UPDATE_DATA_V1_2,
-    'duration': 15,
-    'is_enabled': True,
-    'nocache': False,
-    'skip_asset_check': False,
-}
-
-parametrize_version = parametrize(
-    'version',
-    [('v1',), ('v1_1',), ('v1_2',), ('v2',)],
-)
-
-
-class CRUDAssetEndpointsTest(TestCase, ParametrizedTestCase):
-    def setUp(self):
-        self.client = APIClient()
-
-    def get_assets(self, version):
-        asset_list_url = reverse(f'api:asset_list_{version}')
-        response = self.client.get(asset_list_url)
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        return response.data
-
-    def get_request_data(self, data, version):
-        if version in ['v1', 'v1_1']:
-            return {
-                'model': json.dumps(data)
-            }
-        else:
-            return data
-
-    def create_asset(self, data, version):
-        asset_list_url = reverse(f'api:asset_list_{version}')
-        return self.client.post(
-            asset_list_url,
-            data=self.get_request_data(data, version)
-        ).data
-
-    def update_asset(self, asset_id, data, version):
-        return self.client.put(
-            reverse(f'api:asset_detail_{version}', args=[asset_id]),
-            data=self.get_request_data(data, version)
-        ).data
-
-    def get_asset(self, asset_id, version):
-        url = reverse(f'api:asset_detail_{version}', args=[asset_id])
-        return self.client.get(url).data
-
-    def delete_asset(self, asset_id, version):
-        url = reverse(f'api:asset_detail_{version}', args=[asset_id])
-        return self.client.delete(url)
-
-    @parametrize_version
-    def test_get_assets_when_first_time_setup_should_initially_return_empty(self, version):  # noqa: E501
-        asset_list_url = reverse(f'api:asset_list_{version}')
-        response = self.client.get(asset_list_url)
-        assets = response.data
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(assets), 0)
-
-    @parametrize_version
-    def test_create_asset_should_return_201(self, version):
-        asset_list_url = reverse(f'api:asset_list_{version}')
-        response = self.client.post(
-            asset_list_url,
-            data=self.get_request_data(ASSET_CREATION_DATA, version)
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-
-        self.assertEqual(response.data['name'], 'Anthias')
-        self.assertEqual(response.data['uri'], 'https://anthias.screenly.io')
-        self.assertEqual(response.data['is_enabled'], 0)
-        self.assertEqual(response.data['nocache'], 0)
-        self.assertEqual(response.data['play_order'], 0)
-        self.assertEqual(response.data['skip_asset_check'], 0)
-
-    @parametrize_version
-    def test_get_assets_after_create_should_return_1_asset(self, version):
-        self.create_asset(ASSET_CREATION_DATA, version)
-
-        assets = self.get_assets(version)
-        self.assertEqual(len(assets), 1)
-
-    @parametrize_version
-    def test_get_asset_by_id_should_return_asset(self, version):
-        expected_asset = self.create_asset(ASSET_CREATION_DATA, version)
-        asset_id = expected_asset['asset_id']
-        actual_asset = self.get_asset(asset_id, version)
-
-        self.assertEqual(expected_asset, actual_asset)
-
-    @parametrize_version
-    def test_update_asset_should_return_updated_asset(self, version):
-        expected_asset = self.create_asset(ASSET_CREATION_DATA, version)
-        asset_id = expected_asset['asset_id']
-
-        if version == 'v2':
-            data = ASSET_UPDATE_DATA_V2
-        else:
-            data = ASSET_UPDATE_DATA_V1_2
-
-        updated_asset = self.update_asset(
-            asset_id,
-            data=data,
-            version=version,
-        )
-
-        self.assertEqual(updated_asset['name'], 'Anthias')
-        self.assertEqual(updated_asset['uri'], 'https://anthias.screenly.io')
-        self.assertEqual(updated_asset['duration'], data['duration'])
-        self.assertEqual(updated_asset['is_enabled'], data['is_enabled'])
-        self.assertEqual(updated_asset['play_order'], data['play_order'])
-
-    @parametrize_version
-    def test_delete_asset_should_return_204(self, version):
-        asset = self.create_asset(ASSET_CREATION_DATA, version)
-        asset_id = asset['asset_id']
-
-        response = self.delete_asset(asset_id, version)
-        assets = self.client.get(ASSET_LIST_V1_1_URL).data
-
-        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
-        self.assertEqual(len(assets), 0)
 
 
 class V1EndpointsTest(TestCase, ParametrizedTestCase):
@@ -196,7 +45,7 @@ class V1EndpointsTest(TestCase, ParametrizedTestCase):
 
     def test_file_asset(self):
         project_base_path = django_settings.BASE_DIR
-        image_path = path.join(
+        image_path = os.path.join(
             project_base_path,
             'static/img/standby.png',
         )
@@ -210,7 +59,7 @@ class V1EndpointsTest(TestCase, ParametrizedTestCase):
         data = response.data
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertTrue(path.exists(data['uri']))
+        self.assertTrue(os.path.exists(data['uri']))
         self.assertEqual(data['ext'], '.png')
 
     def test_playlist_order(self):
